106 refactor ast

+++ src/parser/parser.h
@@ -24,7 +24,7 @@ void fill_if_node(struct ast *if_node, struct ast *node);
 
 void init_redirection_node(struct ast *redirection_node);
 
-void fill_redirection_node(struct ast *redirection_node, size_t position,
+void fill_redirection_node(struct ast *redirection_node, int position,
                            char *str);
 
 void init_pipeline_node(struct ast *if_node);
